release: v5.1.0 - Plan Database and Output Service
- Add Plan Database with SQLite persistent storage (PR #60)
- Add Output Service for structured document generation (PR #62)
- Fix 38 compiler warnings in new features
- Add /plan command (list, status, export, cleanup)
- Add /output command (list, latest, open, delete, size, cleanup)
- Add style command documentation
- Update README with new features and commands

Developed by Roberto D'Angelo with the help of a team of AI agents.

+++ src/orchestrator/planning.c
@@ -115,7 +115,7 @@ bool orch_plan_get_progress(ExecutionPlan* plan, int* total, int* completed, flo
         if (plan_db_get_progress(plan->db_id, &progress) == PLAN_DB_OK) {
             if (total) *total = progress.total;
             if (completed) *completed = progress.completed;
-            if (percent) *percent = progress.percent_complete;
+            if (percent) *percent = (float)progress.percent_complete;
             return true;
         }
     }
